Adjsutet for desing pattern
