+++ src/hiof/frameworks/group15/HolderInterface.java
@@ -1,0 +1,8 @@
+package hiof.frameworks.group15;
+
+import java.io.IOException;
+
+public interface HolderInterface {
+    public void generateFile(String filename, String info) throws IOException;
+
+}
